requirements.txt
fastapi
uvicorn
requests
feedparser
openai
google-generativeai
fastapi
uvicorn
feedparser
requests
google-generativeai
fastapi
uvicorn
requests
feedparser
python-dotenv
schedule
pillow